issues

+++ server/jagarv/src/main/java/com/app/jagarv/dto/product/read/BestSellerDTO.java
@@ -41,19 +41,19 @@ public class BestSellerDTO {
         this.pictures = pictures;
     }
 
-    public Integer getStock() {
+    public Long getStock() {
         return stock;
     }
 
-    public void setStock(Integer stock) {
+    public void setStock(Long stock) {
         this.stock = stock;
     }
 
-    public Double getPrice() {
+    public Long getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(Long price) {
         this.price = price;
     }
